runtime: shuffle stuff
revert this?
